revert unnecessary changes
