+++ .github/workflows/test_onnxruntime.yml
@@ -17,8 +17,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-20.04, windows-2019, macos-15]
         transformers-version: ["latest"]
+        os: [ubuntu-20.04, windows-2019, macos-13]
         include:
           - transformers-version: "4.45.*"
             os: ubuntu-20.04
